scripts: hid_configurator: Move LED stream to separate files
Change moves LED stream related code to separate files.

Jira:DESK-593
